Refactoring: Create a GraphQLCommon class
